+++ app/src/main/java/com/android/roomdbtest/presentation/home_screen/HomeScreen.kt
@@ -25,68 +25,3 @@ fun HomeScreen(noteViewModel: UpdateNoteViewModel,navController: NavController) 
     val notes = noteViewModel.noteList.collectAsState().value
     NoteListScreen(notes = notes, noteViewModel =noteViewModel, navController = navController )
 }
-
-@OptIn(ExperimentalMaterialApi::class)
-@SuppressLint("UnusedMaterialScaffoldPaddingParameter")
-@Composable
-fun NoteListScreen(
-    notes: List<Note>,
-    noteViewModel: UpdateNoteViewModel,
-    navController: NavController
-) {
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                title = { Text(text = "Notes") }
-            )
-        },
-        content = {
-            Column(
-                modifier = Modifier
-                    .padding(16.dp)
-                    .fillMaxSize()
-            ) {
-                LazyColumn(
-                    modifier = Modifier.weight(1f)
-                ) {
-                    items(notes) { note ->
-                        Card(
-                            modifier = Modifier
-                                .padding(8.dp)
-                                .fillMaxWidth(),
-                            elevation = 8.dp,
-                            shape = RoundedCornerShape(16.dp),
-                            onClick = {  }
-                        ) {
-                            Column(
-                                modifier = Modifier.padding(16.dp)
-                            ) {
-                                Text(text = note.title, style = MaterialTheme.typography.h6)
-                                Spacer(modifier = Modifier.height(8.dp))
-                                Text(text = note.content, style = MaterialTheme.typography.body2)
-                                Spacer(modifier = Modifier.height(8.dp))
-                                Row(
-                                    modifier = Modifier.fillMaxWidth(),
-                                    horizontalArrangement = Arrangement.End
-                                ) {
-                                    IconButton(
-                                        onClick = { noteViewModel.deleteNote(note = note) }
-                                    ) {
-                                        Icon(Icons.Filled.Delete, contentDescription = "Delete")
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-                FloatingActionButton(
-                    onClick = { navController.navigate(Screens.UpdateNoteScreen.route)},
-                    modifier = Modifier.align(Alignment.End),
-                    content = {
-                        Icon(Icons.Filled.Add, contentDescription = "Add")
-                    }
-                )
-            }
-        }
-    )
-}
