+++ app/src/main/java/com/android/roomdbtest/presentation/update_note/UpdateNoteScreen.kt
@@ -3,12 +3,16 @@ package com.android.roomdbtest.presentation.update_note
 import android.annotation.SuppressLint
 import android.widget.Toast
 import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.input.KeyboardCapitalization
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.android.roomdbtest.domain.model.Note
@@ -48,13 +52,17 @@ fun UpdateNoteScreen(
                 OutlinedTextField(
                     value = title,
                     onValueChange = { title = it },
-                    label = { Text(text = "Note Title") }
+                    label = { Text(text = "Note Title") },
+                    singleLine = true,
+                    keyboardOptions = KeyboardOptions(capitalization = KeyboardCapitalization.Sentences)
                 )
                 Spacer(modifier = Modifier.height(16.dp))
                 OutlinedTextField(
                     value = content,
                     onValueChange = { content = it },
-                    label = { Text(text = "Note Content") }
+                    label = { Text(text = "Note Content") },
+                    keyboardOptions = KeyboardOptions(capitalization = KeyboardCapitalization.Sentences)
+
                 )
                 Spacer(modifier = Modifier.weight(1f))
                 Button(
@@ -70,7 +78,8 @@ fun UpdateNoteScreen(
                        navController.navigate(Screens.HomeScreen.route)
                               },
                     modifier = Modifier.fillMaxWidth(),
-                    contentPadding = PaddingValues(vertical = 12.dp)
+                    contentPadding = PaddingValues(vertical = 12.dp),
+                    colors = ButtonDefaults.buttonColors(backgroundColor = Color(0xFF576CBC))
                 ) {
                     Text(text = "Save")
                 }
